+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes will be documented in this file.
 
 ## Head
 
+## 0.7.1 &ndash; 2021-05-30
+
 ### Changed
 
 * Dropped requirement for calling `Logger::initialize(...)`